Splitting light into SSLightBase and SSDirectionalLight
