+++ SimpleScene/_SimpleScene.csproj
@@ -72,7 +72,6 @@
     <Compile Include="Cameras\SSCameraThirdPerson.cs" />
     <Compile Include="Core\SSRay.cs" />
     <Compile Include="Core\SSTexture.cs" />
-    <Compile Include="Lights\SSLight.cs" />
     <Compile Include="Meshes\SSAbstractMesh.cs" />
     <Compile Include="Meshes\SSIndexBuffer.cs" />
     <Compile Include="Meshes\SSVertexBuffer.cs" />
@@ -113,6 +112,8 @@
     <Compile Include="Lights\SSParallelSplitShadowMap.cs" />
     <Compile Include="Core\SSPssmShaderProgram.cs" />
     <Compile Include="Lights\SSShadowMapBase.cs" />
+    <Compile Include="Lights\SSLightBase.cs" />
+    <Compile Include="Lights\SSDirectionalLight.cs" />
   </ItemGroup>
   <Import Project="$(MSBuildBinPath)\Microsoft.CSharp.targets" />
   <ItemGroup />
